Fixed compilation
