+++ taichi/common/virtual_dir.cpp
