+++ taichi/util/CMakeLists.txt
@@ -14,7 +14,6 @@ target_sources(taichi_util
     str.cpp
     testing.cpp
     zip.cpp
-    virtual_dir.cpp
   )
 
 target_include_directories(taichi_util
