+++ taichi/util/zip.cpp
@@ -1,4 +1,4 @@
-#include "zip.h"
+#include "taichi/common/core.h"
 
 #if defined(__GNUC__)
 // Ensure we get the 64-bit variants of the CRT's file I/O calls
@@ -108,45 +108,6 @@ std::vector<uint8> read(const std::string fn, bool verbose) {
   std::vector<uint8> ret(p, p + file_stat.m_uncomp_size);
   mz_free((void *)p);
   return ret;
-}
-
-bool ZipArchive::try_from_bytes(const void *data, size_t size, ZipArchive &ar) {
-  mz_bool succ = MZ_TRUE;
-  ar.file_dict.clear();
-
-  mz_zip_archive zip;
-  mz_zip_zero_struct(&zip);
-
-  {
-    succ &= mz_zip_reader_init_mem(&zip, data, size, 0);
-    if (succ != MZ_TRUE) {
-      goto fail;
-    }
-  }
-
-  {
-    mz_uint nfile = mz_zip_reader_get_num_files(&zip);
-    for (mz_uint i = 0; i < nfile; ++i) {
-      mz_zip_archive_file_stat file_stat;
-      succ &= mz_zip_reader_file_stat(&zip, i, &file_stat);
-      if (succ != MZ_TRUE) {
-        goto fail;
-      }
-
-      std::vector<uint8_t> file_data(file_stat.m_uncomp_size);
-      succ &= mz_zip_reader_extract_to_mem(&zip, i, file_data.data(),
-                                           file_data.size(), 0);
-      if (succ != MZ_TRUE) {
-        goto fail;
-      }
-      ar.file_dict[file_stat.m_filename] = std::move(file_data);
-    }
-  }
-
-fail:
-  succ &= mz_zip_reader_end(&zip);
-
-  return succ;
 }
 
 }  // namespace zip
